+++ pytorch_lightning/metrics/__init__.py
@@ -15,7 +15,7 @@ from pytorch_lightning.metrics.classification import (
     Precision,
     PrecisionRecall,
 )
-from pytorch_lightning.metrics.sk_metrics import (
+from pytorch_lightning.metrics.sklearns import (
     AUC,
     PrecisionRecallCurve,
     SklearnMetric,
